+++ examples/angular2/app/src/RxDB.d.ts
@@ -3,25 +3,13 @@
  * @type {[type]}
  */
 
-import * as RxDB from 'rxdb';
+import { RxDocument, RxCollection, RxDatabase } from 'rxdb';
 import { Observable } from 'rxjs';
 
-declare interface RxHeroDocumentData {
+declare interface RxHeroDocumentType {
     name?: string;
     color?: string;
     maxHP?: number;
-    hp?: number;
-    team?: string;
-    skills?: Array<{
-        name?: string,
-        damage?: string
-    }>;
-}
-
-export class RxHeroDocument extends RxDB.RxDocument {
-    name: string;
-    color: string;
-    maxHP: number;
     hp?: number;
     team?: string;
     skills?: Array<{
@@ -33,15 +21,16 @@ export class RxHeroDocument extends RxDB.RxDocument {
     hpPercent(): number;
 }
 
-declare class RxHeroCollection extends RxDB.RxCollection<RxHeroDocument> {
+export type RxHeroDocument = RxDocument<RxHeroDocumentType>;
+
+declare class RxHeroCollection extends RxCollection<RxHeroDocumentType> {
 }
 
-export class RxHeroesDatabase extends RxDB.RxDatabase {
+export class RxHeroesDatabase extends RxDatabase {
     hero?: RxHeroCollection;
 }
 
 export default {
-    RxHeroDocument,
     RxHeroCollection,
     RxHeroesDatabase
 };