+++ plugins/core/index.d.ts
@@ -1,0 +1,2 @@
+import * as RxDB from '../../src/index';
+export default RxDB;