+++ src/plugins/validate.js
@@ -40,7 +40,7 @@ const _getValidator = function(schemaPath = '') {
 };
 
 /**
- * validates the given object agains the schema
+ * validates the given object against the schema
  * @param  {any} obj
  * @param  {String} [schemaPath=''] if given, the sub-schema will be validated
  * @throws {RxError} if not valid