+++ CHANGELOG.md
@@ -3,15 +3,18 @@
 ## 6.0.2 (coming soon)
 
 Features:
+    - Added [attachments](https://pubkey.github.io/rxdb/rx-attachment.html)
     - Added [final fields](https://pubkey.github.io/rxdb/rx-schema.html#final)
 
 Bugfixes:
     - Removed type `ingoreDuplicate` [#314](https://github.com/pubkey/rxdb/issues/314)
     - Added error-message when you json-import on a non-existing collection [#319](https://github.com/pubkey/rxdb/issues/319)
+    - Allow windows-foldernames (with backslash) as collection-name [343](https://github.com/pubkey/rxdb/issues/343)
 
 Other:
     - Split out idle-queue into own [npm-module](http://npmjs.com/package/custom-idle-queue)
     - Enfore usage of strict-equality via eslint
+    - Improved typings [#329](https://github.com/pubkey/rxdb/pull/329) by @ihadeed
 
 ## 6.0.1 (September 20, 2017)